fix: undefined `global` in some cjs environments
Having a variable named `global` has been something of a de-facto
standard for CommonJS environments and CommonJS bundlers, due to it
being available in Node.js module contexts. However, now there is at
least one popular CommonJS environment that does not define it.

When running in Node.js, we want to make use of the `process.nextTick`
function for scheduling callbacks. Instead of checking if
`global.process` exists, this fix uses the old but kinda gangly
`typeof process === 'object` approach to avoid needing to refer to a
`global`/`globalThis`/`self`/`window` object.

Fixes global is undefined  #7.